separate static and dynamic head tags
