+++ front/main/app/mixins/head-tags-dynamic.js
@@ -10,7 +10,6 @@ export default Ember.Mixin.create({
 	afterModel(model) {
 		this._super(...arguments);
 
-		this.setStaticHeadTags();
 		this.setDynamicHeadTags(model);
 		this.removeServerTags();
 	},
@@ -22,37 +21,12 @@ export default Ember.Mixin.create({
 	 * @returns {void}
 	 */
 	removeServerTags() {
-		if (!this.controllerFor('application').get('serverTagsRemoved')) {
+		const headData = this.get('headData');
+
+		if (!headData.get('serverTagsRemoved')) {
 			Ember.$('[data-server-head-tags]').remove();
-			this.controllerFor('application').set('serverTagsRemoved', true);
+			headData.set('serverTagsRemoved', true);
 		}
-	},
-
-	/**
-	 * This function sets static head tags defined in templates/head.hbs
-	 * This is for head tags which are set only once
-	 *
-	 * @returns {void}
-	 */
-	setStaticHeadTags() {
-		const wikiVariables = Ember.get(Mercury, 'wiki'),
-			verticalColors = {
-				comics: '#ff5400',
-				games: '#94d11f',
-				books: '#ff7f26',
-				movies: '#09d3bf',
-				lifestyle: '#ffd000',
-				music: '#c819ad',
-				tv: '#00b7e0'
-			};
-
-		this.get('headData').setProperties({
-			favicon: wikiVariables.favicon,
-			themeColor: verticalColors[wikiVariables.vertical],
-			gaUrl: M.prop('gaUrl'),
-			qualarooScript: M.prop('qualarooScript'),
-			optimizelyScript: M.prop('optimizelyScript')
-		});
 	},
 
 	/**
