+++ front/main/app/routes/main-page-category.js
@@ -2,9 +2,9 @@ import Ember from 'ember';
 import MainPageRouteMixin from '../mixins/main-page-route';
 import RouteWithAdsMixin from '../mixins/route-with-ads';
 import CuratedContentModel from '../models/curated-content';
-import HeadTagsMixin from '../mixins/head-tags';
+import HeadTagsDynamicMixin from '../mixins/head-tags-dynamic';
 
-export default Ember.Route.extend(MainPageRouteMixin, HeadTagsMixin, RouteWithAdsMixin, {
+export default Ember.Route.extend(MainPageRouteMixin, HeadTagsDynamicMixin, RouteWithAdsMixin, {
 	/**
 	 * @param {*} params
 	 * @returns {Ember.RSVP.Promise}
