+++ front/main/app/routes/wiki-page.js
@@ -3,13 +3,13 @@ import ArticleHandler from '../utils/wiki-handlers/article';
 import CategoryHandler from '../utils/wiki-handlers/category';
 import CuratedMainPageHandler from '../utils/wiki-handlers/curated-main-page';
 import RouteWithAdsMixin from '../mixins/route-with-ads';
-import HeadTagsMixin from '../mixins/head-tags';
+import HeadTagsDynamicMixin from '../mixins/head-tags-dynamic';
 import getPageModel from '../utils/wiki-handlers/wiki-page';
 import {normalizeToUnderscore} from 'common/utils/string';
 import {setTrackContext, trackPageView} from 'common/utils/track';
 import {namespace as MediawikiNamespace, isContentNamespace} from '../utils/mediawiki-namespace';
 
-export default Ember.Route.extend(RouteWithAdsMixin, HeadTagsMixin, {
+export default Ember.Route.extend(RouteWithAdsMixin, HeadTagsDynamicMixin, {
 	redirectEmptyTarget: false,
 	wikiHandler: null,
 	currentUser: Ember.inject.service(),
